chore: clean up TimerScreen into smaller components
